Move KarmaFileWatcher to JsTestAdapter
