+++ JsTestAdapter/JsTestAdapter.csproj
@@ -117,6 +117,7 @@
   <ItemGroup>
     <Compile Include="Helpers\ExceptionExtensions.cs" />
     <Compile Include="Helpers\Json.cs" />
+    <Compile Include="Helpers\FileWatcher.cs" />
     <Compile Include="Helpers\PathUtils.cs" />
     <Compile Include="Helpers\ProjectListener.cs" />
     <Compile Include="Helpers\Sha1Utils.cs" />
