+++ KarmaTestAdapter/KarmaTestAdapter.csproj
@@ -88,7 +88,6 @@
     <Compile Include="Logging\KarmaLogger.cs" />
     <Compile Include="Logging\KarmaServerTestLogger.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestAdapter\KarmaFileWatcher.cs" />
     <Compile Include="TestAdapter\KarmaSettings.cs" />
     <Compile Include="TestAdapter\KarmaTestContainer.cs" />
     <Compile Include="TestAdapter\KarmaTestContainerBase.cs" />
